fix: register fvcom_minimal_dataset in conftest for CI Share the FVCOM synthetic fixture via tests/conftest.py so the integration ctl-write test can find it; run pre-push ci-local in ofs_dps.
