+++ tests/write_ofs_ctlfile_minimal_intake_test.py
@@ -31,132 +31,10 @@ import logging
 import os
 from types import SimpleNamespace
 
-import numpy as np
 import pytest
-import xarray as xr
 
 from ofs_skill.model_processing.write_ofs_ctlfile import write_ofs_ctlfile
-from ofs_skill.utils.file_headers import OBS_CTL_HEADER
-
-# ---------------------------------------------------------------------------
-# Fixture: synthetic multi-file FVCOM stations dataset
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture()
-def fvcom_minimal_dataset(tmp_path):
-    """Build two small FVCOM stations files + combine via xr.open_mfdataset.
-
-    With ``data_vars='minimal'`` and a single time-concat dim, static
-    vars (``lon``, ``lat``, ``h``, ``siglay``) come back at their native
-    shape (no time replication). ``zeta`` is the only time-varying var
-    here.
-
-    Returns
-    -------
-    (xr.Dataset, np.ndarray, np.ndarray)
-        The combined dataset and the original 1-D lon / lat arrays the
-        caller can use to verify the written ctl file.
-    """
-    n_station = 6
-    n_siglay = 3
-    n_time = 4
-
-    lon_1d = np.linspace(-71.0, -68.0, n_station, dtype=np.float64)
-    lat_1d = np.linspace(41.0, 44.0, n_station, dtype=np.float64)
-    h_1d = np.linspace(5.0, 30.0, n_station, dtype=np.float64)
-    siglay = np.tile(
-        np.linspace(-1.0, 0.0, n_siglay, dtype=np.float64)[:, None],
-        (1, n_station),
-    )
-
-    def _make_file(path, t_offset):
-        ds = xr.Dataset(
-            data_vars={
-                'lon': (('station',), lon_1d),
-                'lat': (('station',), lat_1d),
-                'h': (('station',), h_1d),
-                'siglay': (('siglay', 'station'), siglay),
-                'zeta': (
-                    ('time', 'station'),
-                    np.zeros((n_time, n_station), dtype=np.float64),
-                ),
-            },
-            coords={
-                'time': (
-                    np.datetime64('2026-02-16T00')
-                    + (t_offset + np.arange(n_time)) * np.timedelta64(1, 'h')
-                ),
-            },
-        )
-        ds.to_netcdf(path)
-        ds.close()
-
-    f1 = tmp_path / 'fvcom_stations_a.nc'
-    f2 = tmp_path / 'fvcom_stations_b.nc'
-    _make_file(f1, 0)
-    _make_file(f2, n_time)
-
-    combined = xr.open_mfdataset(
-        [str(f1), str(f2)],
-        data_vars='minimal',
-        combine='nested',
-        concat_dim='time',
-    )
-
-    # Pre-condition: static coords MUST remain 1-D under 'minimal'.
-    # If a future xarray bump changes that, the test fails loudly here
-    # rather than surfacing as a misleading downstream assertion.
-    assert combined['lon'].dims == ('station',), combined['lon'].dims
-    assert combined['lat'].dims == ('station',), combined['lat'].dims
-
-    return combined, lon_1d, lat_1d
-
-
-# ---------------------------------------------------------------------------
-# Fixture: minimal config file + prop / extract / station_ctl files
-# ---------------------------------------------------------------------------
-
-
-def _write_minimal_config(tmp_path):
-    """Write a minimal INI config the writer needs for read_config_section.
-
-    ``write_ofs_ctlfile`` reads the ``directories`` section to assemble
-    ``prop.model_path`` even though the path is not actually used when
-    only the ctl writer branch runs end-to-end. We supply just enough
-    keys to make ``configparser`` happy.
-    """
-    cfg_path = tmp_path / 'ofs_dps.conf'
-    cfg_path.write_text(
-        '[directories]\n'
-        f'home={tmp_path.as_posix()}\n'
-        'model_historical_dir=%(home)s/example_data\n'
-        'netcdf_dir=netcdf\n'
-    )
-    return cfg_path
-
-
-def _write_obs_station_ctl(control_dir, ofs, name_var, stations):
-    """Write a minimal obs station.ctl file the writer will read.
-
-    Format (per station_ctl_file_extract):
-        <Header Line 1>
-        <Header Line 2>
-        <ID> <ID>_<SRC> "<name>"
-          <lat> <lon> <depth> <pad> <datum>
-    """
-    path = control_dir / f'{ofs}_{name_var}_station.ctl'
-    lines = []
-    for sid, lat, lon, depth in stations:
-        lines.append(f'{sid} {sid}_COOPS "Station {sid}"')
-        lines.append(f'  {lat} {lon} {depth} 0.0 MLLW')
-    path.write_text(OBS_CTL_HEADER + '\n'.join(lines) + '\n')
-    return path
-
-
-# ---------------------------------------------------------------------------
-# Test
-# ---------------------------------------------------------------------------
+from tests.helpers.fvcom_minimal import write_minimal_config, write_obs_station_ctl
 
 
 @pytest.mark.integration
@@ -166,7 +44,7 @@ def test_write_ofs_ctlfile_fvcom_stations_minimal_intake(
     """End-to-end smoke test: writer must not raise on 1-D static coords."""
     combined, lon_1d, lat_1d = fvcom_minimal_dataset
 
-    cfg_path = _write_minimal_config(tmp_path)
+    cfg_path = write_minimal_config(tmp_path)
     control_dir = tmp_path / 'control_files'
     control_dir.mkdir()
 
@@ -178,7 +56,7 @@ def test_write_ofs_ctlfile_fvcom_stations_minimal_intake(
         (f'9000{ni:02d}', float(lat_1d[ni]), float(lon_1d[ni]), 0.0)
         for ni in chosen_nodes
     ]
-    _write_obs_station_ctl(control_dir, 'tbofs', 'wl', stations)
+    write_obs_station_ctl(control_dir, 'tbofs', 'wl', stations)
 
     prop = SimpleNamespace(
         config_file=str(cfg_path),
